+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 1.1.0
 * feat: Rename `list` commands to `ls`.
 * feat: Add `lambda-ai instance-types --available-only` to only list available instance types.
+* feat: Add `lambda-ai instance-types --cheapest` to list the cheapest instance type.
 
 # 1.0.0
 * feat: Initial release.